Move Job behavior into concerns. Only support a single perform method, remove 'perform_method' attribute. Remove backward compatibility code.
