+++ lib/rocket_job/version.rb
@@ -1,4 +1,4 @@
 # encoding: UTF-8
 module RocketJob #:nodoc
-  VERSION = '1.4.0'
+  VERSION = '2.0.0.alpha'
 end
